requirements.txt
flask
pillow
python-memcached